requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pynacl>=1.5.0
requests>=2.0.0
